+++ platforms/android/contribute/src/main/java/pulse/contribute/android/util/Helper.kt
@@ -1,6 +1,5 @@
 package pulse.contribute.android.util
 
-import pulse.contribute.android.ApiInterface
 import java.lang.Exception
 
 class Helper(){
